automatically expand commands
"tbxmanager sh" will now expand to "tbxmanager show". If the expansion is
ambiguous, we give the list of possible choices and break with an error.

The expansion is carried out by tbx_expandChoice. It takes a string and a
cell array of possible choices, and return the choice which starts with the
string.