better check of whether the source XML is compatible
We now have the "max_xml_version" option in tbx_setup(), which denotes the
maximal version of source XML files compatible with this particular version
of tbxmanager.

Then a source XML file is declared compatible is its version number is
smaller than or equal to max_xml_version.

Note that the version numbers in the XML files (<tbxmanager version="x.y">)
must be in numerical format, e.g. "1.0", "2.9", etc.