basic tests
Call tbx_runTests to execute all tests.

A test is designated as "ok" if its output printed to the screen matches the
content of the corresponding ".out" file (after removing all blank
characters including spaces, and after replacing @current_date@ by the
current date).

If the test's output differs from the expected outcome, but the test does
not break with an error, it is designated as "wrong".

If the test breaks with an error, it is designated as "error".